+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Get catchment tile number from the input restart file instead of input bcs which may be empty
 - Speedup remap_restarts.py package by adding more processes to catch tiles for fine resolutions.
 
 ### Removed